# api-spec/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSubQuoteInput": {
        "additionalProperties": false,
        "properties": {
          "contractorEmail": {
            "nullable": true,
            "type": "string"
          },
          "contractorName": {
            "nullable": true,
            "type": "string"
          },
          "currency": {
            "nullable": true,
            "type": "string"
          },
          "projectId": {
            "format": "int32",
            "type": "integer"
          },
          "scope": {
            "nullable": true,
            "type": "string"
          },
          "trade": {
            "nullable": true,
            "type": "string"
          },
          "validDays": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PortalSubmitInput": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "format": "double",
            "type": "number"
          },
          "notes": {
            "nullable": true,
            "type": "string"
          },
          "respondentName": {
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "QuoteInput": {
        "additionalProperties": false,
        "properties": {
          "attachmentUrl": {
            "nullable": true,
            "type": "string"
          },
          "currency": {
            "nullable": true,
            "type": "string"
          },
          "note": {
            "nullable": true,
            "type": "string"
          },
          "price": {
            "format": "double",
            "type": "number"
          },
          "quotedOn": {
            "format": "date",
            "type": "string"
          },
          "resourceId": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "resourceType": {
            "nullable": true,
            "type": "string"
          },
          "supplier": {
            "nullable": true,
            "type": "string"
          },
          "unit": {
            "nullable": true,
            "type": "string"
          },
          "validUntil": {
            "format": "date",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SubQuoteDecisionInput": {
        "additionalProperties": false,
        "properties": {
          "accept": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SubcontractorInput": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "nullable": true,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "name": {
            "nullable": true,
            "type": "string"
          },
          "unit": {
            "nullable": true,
            "type": "string"
          },
          "unitRate": {
            "format": "double",
            "type": "number"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "description": "JWT Authorization header. Example: \"Bearer {token}\"",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Multi-tenant construction bid-estimating API. The generated TS client (lib/api-generated.d.ts) is produced from this spec; a runtime-vs-checked-in diff in CI fails the build on contract drift (19.5).",
    "title": "BidBuilder API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/portal/{token}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalSubmitInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
...
    "/api/quotes": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "resourceId",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
    "/api/quotes/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
...
    "/api/resources/subcontractors": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubcontractorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
    "/api/resources/subcontractors/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubcontractorInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
...
    "/api/subcontractor-quotes": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubQuoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
    "/api/subcontractor-quotes/{id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
    "/api/subcontractor-quotes/{id}/decision": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubQuoteDecisionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "BidBuilder.Api"
        ]
      }
    },
...
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}